pyproject.toml
[tool.poetry]
name = "pgoops"
version = "0.1.0"
description = "pgoops"
authors = ["闫世成 <yans121@sina.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
django = "^3.2.13"
djangorestframework-jwt = "^1.11.0"
djangorestframework = "^3.13.1"
django-filter = "^21.1"
casbin = "^1.15.4"
pymysql = "^1.0.2"
celery = "^5.2.6"
django-celery-beat = "^2.2.1"
simplejson = "^3.17.6"
django-cors-headers = "^3.11.0"
django-mptt = "^0.13.4"
dictdiffer = "^0.9.0"
pillow = "^9.1.0"
ansible-runner = "^2.1.3"
redis = "^4.2.2"
channels = "^3.0.4"
channels-redis = "^3.4.0"
django-redis = "^5.2.0"
requests = "^2.27.1"
eventlet = "^0.33.0"
uvicorn = "^0.17.6"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[[tool.poetry.source]]
name = "douban"
url = "https://pypi.doubanio.com/simple/"
default = false